Add PatientRegisterViewHolder tests
