+++ rdt-open-reader/src/test/java/io/ona/rdt/presenter/PatientProfileActivityPresenterTest.java
@@ -1,8 +1,7 @@
-package io.ona.rdt.robolectric.presenter;
-
-import android.app.Activity;
+package io.ona.rdt.presenter;
 
 import org.json.JSONException;
+import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -17,13 +16,11 @@ import org.powermock.reflect.Whitebox;
 import org.smartregister.repository.EventClientRepository;
 
 import io.ona.rdt.application.RDTApplication;
+import io.ona.rdt.callback.OnFormSavedCallback;
 import io.ona.rdt.contract.PatientProfileActivityContract;
-import io.ona.rdt.contract.PatientProfileFragmentContract;
-import io.ona.rdt.domain.Patient;
-import io.ona.rdt.interactor.PatientProfileFragmentInteractor;
-import io.ona.rdt.presenter.PatientProfileFragmentPresenter;
+import io.ona.rdt.interactor.PatientProfileActivityInteractor;
+import io.ona.rdt.presenter.PatientProfileActivityPresenter;
 
-import static io.ona.rdt.util.Constants.Form.RDT_TEST_FORM;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
@@ -36,12 +33,12 @@ import static org.powermock.api.mockito.PowerMockito.mockStatic;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({RDTApplication.class})
-public class PatientProfileFragmentPresenterTest {
+public class PatientProfileActivityPresenterTest {
 
     @Mock
-    private PatientProfileFragmentInteractor interactor;
+    private PatientProfileActivityInteractor interactor;
     @Mock
-    PatientProfileFragmentContract.View fragment;
+    PatientProfileActivityContract.View activity;
     @Mock
     private RDTApplication rdtApplication;
     @Mock
@@ -50,27 +47,26 @@ public class PatientProfileFragmentPresenterTest {
     protected EventClientRepository eventClientRepository;
 
     @Captor
-    private ArgumentCaptor<Patient> patientArgumentCaptor;
+    private ArgumentCaptor<JSONObject> jsonObjectCaptor;
 
-    private PatientProfileFragmentPresenter presenter;
+    private PatientProfileActivityPresenter presenter;
 
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
         mockStaticMethods();
-        presenter = new PatientProfileFragmentPresenter(fragment);
+        presenter = new PatientProfileActivityPresenter(activity);
     }
 
     @Test
-    public void testLaunchFormShouldLaunchRDTTestForm() throws JSONException {
-        Whitebox.setInternalState(presenter, "patientProfileFragmentInteractor", interactor);
-        Activity activity = mock(Activity.class);
-        Patient patient = new Patient("name", "sex", "entity_id", "patient_id");
-        presenter.launchForm(activity, patient);
-        verify(interactor).launchForm(eq(activity), eq(RDT_TEST_FORM), patientArgumentCaptor.capture());
-
-        Patient actualPatient = patientArgumentCaptor.getValue();
-        assertEquals(patient, actualPatient);
+    public void testSaveFormShouldSaveForm() throws JSONException {
+        OnFormSavedCallback callback = mock(OnFormSavedCallback.class);
+        JSONObject jsonForm = new JSONObject();
+        jsonForm.put("key", "value");
+        Whitebox.setInternalState(presenter, "interactor", interactor);
+        presenter.saveForm(jsonForm.toString(), callback);
+        verify(interactor).saveForm(jsonObjectCaptor.capture(), eq(callback));
+        assertEquals(jsonForm.toString(), jsonObjectCaptor.getValue().toString());
     }
 
     private void mockStaticMethods() {
