+++ rdt-open-reader/src/test/java/io/ona/rdt/presenter/PatientProfileFragmentPresenterTest.java
@@ -1,7 +1,8 @@
-package io.ona.rdt.robolectric.presenter;
+package io.ona.rdt.presenter;
+
+import android.app.Activity;
 
 import org.json.JSONException;
-import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -16,11 +17,13 @@ import org.powermock.reflect.Whitebox;
 import org.smartregister.repository.EventClientRepository;
 
 import io.ona.rdt.application.RDTApplication;
-import io.ona.rdt.callback.OnFormSavedCallback;
 import io.ona.rdt.contract.PatientProfileActivityContract;
-import io.ona.rdt.interactor.PatientProfileActivityInteractor;
-import io.ona.rdt.presenter.PatientProfileActivityPresenter;
+import io.ona.rdt.contract.PatientProfileFragmentContract;
+import io.ona.rdt.domain.Patient;
+import io.ona.rdt.interactor.PatientProfileFragmentInteractor;
+import io.ona.rdt.presenter.PatientProfileFragmentPresenter;
 
+import static io.ona.rdt.util.Constants.Form.RDT_TEST_FORM;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
@@ -33,12 +36,12 @@ import static org.powermock.api.mockito.PowerMockito.mockStatic;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({RDTApplication.class})
-public class PatientProfileActivityPresenterTest {
+public class PatientProfileFragmentPresenterTest {
 
     @Mock
-    private PatientProfileActivityInteractor interactor;
+    private PatientProfileFragmentInteractor interactor;
     @Mock
-    PatientProfileActivityContract.View activity;
+    PatientProfileFragmentContract.View fragment;
     @Mock
     private RDTApplication rdtApplication;
     @Mock
@@ -47,26 +50,27 @@ public class PatientProfileActivityPresenterTest {
     protected EventClientRepository eventClientRepository;
 
     @Captor
-    private ArgumentCaptor<JSONObject> jsonObjectCaptor;
+    private ArgumentCaptor<Patient> patientArgumentCaptor;
 
-    private PatientProfileActivityPresenter presenter;
+    private PatientProfileFragmentPresenter presenter;
 
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
         mockStaticMethods();
-        presenter = new PatientProfileActivityPresenter(activity);
+        presenter = new PatientProfileFragmentPresenter(fragment);
     }
 
     @Test
-    public void testSaveFormShouldSaveForm() throws JSONException {
-        OnFormSavedCallback callback = mock(OnFormSavedCallback.class);
-        JSONObject jsonForm = new JSONObject();
-        jsonForm.put("key", "value");
-        Whitebox.setInternalState(presenter, "interactor", interactor);
-        presenter.saveForm(jsonForm.toString(), callback);
-        verify(interactor).saveForm(jsonObjectCaptor.capture(), eq(callback));
-        assertEquals(jsonForm.toString(), jsonObjectCaptor.getValue().toString());
+    public void testLaunchFormShouldLaunchRDTTestForm() throws JSONException {
+        Whitebox.setInternalState(presenter, "patientProfileFragmentInteractor", interactor);
+        Activity activity = mock(Activity.class);
+        Patient patient = new Patient("name", "sex", "entity_id", "patient_id");
+        presenter.launchForm(activity, patient);
+        verify(interactor).launchForm(eq(activity), eq(RDT_TEST_FORM), patientArgumentCaptor.capture());
+
+        Patient actualPatient = patientArgumentCaptor.getValue();
+        assertEquals(patient, actualPatient);
     }
 
     private void mockStaticMethods() {
